requirements.txt
fastapi
uvicorn
langchain
chromadb
python-dotenv
openai
